requirements.txt
Django==3.1.1
python-dotenv==0.13.0
djangorestframework~=3.11.1